AIGNITE/AIGNITE/requirements.txt
requests==2.31.0
flask==3.0.0
pytesseract==0.3.10
Pillow==10.0.0
pymupdf==1.24.0
python-docx==1.1.0
chromadb==0.4.15
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0